here we used similarity matrix
there are many technique available to generate extractive summarization. But for simplification i am using unsupervised learning approach to find the similarity in sentences and rank them accordingly. The benefit of using it i don't need to train and build the model we can use directly.
Since we will be representing our sentences as the bunch of vectors, we can use it to find the similarity among sentences. Its measures cosine of the angle between vectors. Angle will be 0 if sentences are similar.